requirements.txt
google-cloud-bigquery
pygal
terminaltables
filecache
click
colorama
